Adding initial directory structure and files (licenses, logic, gnu/linux standards, linguistics)
Added The following directories, which contain initial references:
 - GNU-Linux which contains a Standards directory containing currently the FHS
 - Licenses which contains the CC and GPL directory, each in turn containing major licenses of both FSF and CC. The innner directories include license headers, images/graphics to use, and license in some cases.
 - Logic which contains a PDF file compiled using the create book feature of Wikipedia. The book is a collection of Wikipedia articles on the topic of Logic
 - Linguistics which contains two Wikipedia-compiled books that are collections of WP articles on the topic of Linguistics.

In the future, all books should be uploaded in an editable format I think, but for now, the PDFs will do. When the editable versions of the books and works uploaded here is available, the PDF alternative should be removed (or moved to a release directory or something like that. Haven't figured it out yet)

VC (a.k.a. kOVC, wHVC)